chore: rename folder to match the ts pattern
